+++ bundle-analyzer/src/to-draft-converter.tsx
@@ -1,0 +1,3 @@
+import { toDraft } from 'ricos-content/libs/toDraft';
+
+export default () => toDraft({ nodes: [] });