+++ examples/main/shared/RichContentApp.tsx
@@ -7,9 +7,11 @@ import {
   createWithContent,
   EditorState,
 } from 'wix-rich-content-editor/libs/editorStateConversion';
-import { isSSR, RicosContent, SEOSettings } from 'wix-rich-content-common';
+import { compare, isSSR, RicosContent, SEOSettings } from 'wix-rich-content-common';
 import { getRequestedLocale, normalize } from '../src/utils';
 import { TestAppConfig } from '../src/types';
+import { RichContent } from 'ricos-schema';
+import { fromDraft } from 'ricos-content/libs/migrateSchema';
 
 type Mode = 'demo' | 'test';
 
@@ -28,6 +30,8 @@ interface Props {
 interface State {
   editorState?: EditorState;
   contentState?: RicosContent;
+  content?: RichContent;
+  injectedContent?: RichContent;
   localeResource?: Record<string, string>;
   locale?: string;
   remountKey?: boolean;
@@ -89,6 +93,11 @@ class RichContentApp extends PureComponent<Props, State> {
 
   onRicosEditorChange = (contentState: RicosContent) => this.setState({ contentState });
 
+  onNewContentChange = (content: RichContent) => this.setState({ content });
+
+  onNewInjectedContentChange = (content: RichContent) =>
+    this.setState({ injectedContent: content });
+
   onContentStateChange = (contentState: RicosContent) => {
     this.setState({
       contentState,
@@ -106,7 +115,15 @@ class RichContentApp extends PureComponent<Props, State> {
   };
 
   render() {
-    const { editorState, contentState, localeResource, locale, remountKey } = this.state;
+    const {
+      editorState,
+      contentState,
+      localeResource,
+      locale,
+      remountKey,
+      content,
+      injectedContent,
+    } = this.state;
     const { allLocales, seoMode, isMobile, app: App, testAppConfig } = this.props;
     return (
       <App
@@ -119,6 +136,10 @@ class RichContentApp extends PureComponent<Props, State> {
         localeResource={localeResource}
         onEditorChange={this.onEditorChange}
         onRicosEditorChange={this.onRicosEditorChange}
+        onNewContentChange={this.onNewContentChange}
+        onNewInjectedContentChange={this.onNewInjectedContentChange}
+        content={content}
+        injectedContent={injectedContent}
         onContentStateChange={this.onContentStateChange}
         setLocale={this.setLocaleResource}
         seoMode={seoMode}