+++ packages/ricos-content/web/lib/toDraft.ts
@@ -1,0 +1,7 @@
+export { toDraft, ensureDraftContent } from '../src/migrateSchema/toDraft/toDraft';
+export {
+  convertNodeDataToDraft,
+  convertNodeToDraftData,
+  convertDecorationDataToDraft,
+  convertDecorationToDraftData,
+} from '../src/migrateSchema/toDraft/convertDraftPluginData';